Allow spec simulation steps under cfupd
Previously, we could only do a spec step with an ncfupd, since the invariant
storing the trace of the spec execution was in an ncinv. This has been fine for
previous examples, but the journal requires potentially finishing an Atomically
block during CommitWait if we learn the commit was made durable in the crash
condition.

This commit allows additional steps under a cfupd by transferring the spec
assert to a normal invariant when the crash happens.  Special tokens will be
needed to continue doing transitions at that point of a crash. For the journal
FFI, there is a token per address, and holding any one allows further steps.